pyproject.toml
[project]
name = "linux-deployer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "fastmcp>=2.14.2",
    "psutil>=7.2.1",
    "uvicorn>=0.40.0",
    "python-json-logger>=2.0.7",
    "python-dotenv>=1.0.0",
]
